Add dialog support and basics for entry control.
An Entry is a button but will be rendered differently.  When a button is
clicked, it displays a dialog that contains a keyboard control.  If the
OK button in the dialog is clicked, the text contained in the entry will
be replaced by the text in the keyboard control and a text-changed event
will be raised.

Signed-off-by: Kevin Wells <wellsie1116@gmail.com>